refactor: rename `errorHandler.ts` to `error-handler.ts`
