+++ back-end-challenge-v2-solution/src/app.ts
@@ -1,7 +1,7 @@
 import express, { Application } from "express";
 import { getStatusDB } from "./config/database.js";
 import timeSeriesRoutes from "./routes/time-series.routes.js";
-import { errorHandler } from "./middlewares/errorHandler.js";
+import { errorHandler } from "./middlewares/error-handler.js";
 import AppError from "./errors/app-error.js";
 
 const app: Application = express();
